+++ src/test/java/blackjack/model/deck/PlayingDeckTest.java
@@ -1,11 +1,9 @@
-package blackjack.model;
+package blackjack.model.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.model.card.Card;
-import blackjack.model.deck.DeckGenerator;
-import blackjack.model.deck.PlayingDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
